+++ OMI_code/OMCLDRR/src/cloudret/m_num2string.f90
@@ -22,21 +22,21 @@ function i_string (number, formatstr) result (line)
 ! !ROUTINE:  i_string, etc
 ! 
 ! !DESCRIPTION: similar to IDL "string" function, designed to work
-!		with string interface
+!                with string interface
 !
 ! !CALLING SEQUENCE: 
 !
 !        line = i_string(number, formatstr)
 !     
 ! !INPUT PARAMETERS:   
-integer, dimension(:), intent(in) 		:: number
-!			number    : input number (or array)
-character(len=*), optional, intent(in) 		:: formatstr
-!			formatstr : optional format string
+integer, dimension(:), intent(in)                 :: number
+!                        number    : input number (or array)
+character(len=*), optional, intent(in)                 :: formatstr
+!                        formatstr : optional format string
 !
 ! !OUTPUT PARAMETERS:  
-character(len=72) 				:: line
-!			line      : formatted character string
+character(len=72)                                 :: line
+!                        line      : formatted character string
 !
 ! !SEE ALSO:  IDL documentation
 !
@@ -59,9 +59,9 @@ end function i_string
 function i_string_s (number, formatstr) result (line)
    implicit none
 
-integer, intent(in) 				:: number
-character(len=*), optional, intent(in) 		:: formatstr
-character(len=72) 				:: line
+integer, intent(in)                                 :: number
+character(len=*), optional, intent(in)                 :: formatstr
+character(len=72)                                 :: line
 
 if ( present(formatstr) ) then
   write(line,formatstr) number
@@ -76,9 +76,9 @@ end function i_string_s
 function r_string (number, formatstr) result (line)
    implicit none
 
-real, dimension(:), intent(in) 			:: number
-character(len=*), optional, intent(in)		:: formatstr
-character(len=72) 				:: line
+real (KIND=8), dimension(:), intent(in)                         :: number
+character(len=*), optional, intent(in)                :: formatstr
+character(len=72)                                 :: line
 
 if ( present(formatstr) ) then
   write(line,formatstr) number
@@ -92,9 +92,9 @@ end function r_string
 function r_string_s (number, formatstr) result (line)
    implicit none
 
-real, intent(in)	 			:: number
-character(len=*), optional, intent(in)		:: formatstr
-character(len=72) 				:: line
+real (KIND=8), intent(in)                                 :: number
+character(len=*), optional, intent(in)                :: formatstr
+character(len=72)                                 :: line
 
 if ( present(formatstr) ) then
   write(line,formatstr) number
