+++ OMI_code/OMCLDRR/src/cloudret/mathcons.f90
@@ -11,11 +11,11 @@
 !
       implicit none
 
-      real, parameter ::      pi = 3.14159265359 !   pi
-      real, parameter ::      deg2rad = PI / 180.0  !  convert degrees to radians
-      real, parameter ::      pid2 = PI / 2.0 !   pi/2
-      real, parameter ::      rad2deg = 180.0 / PI !  convert radians to degrees
-      real, parameter ::      twopi = 2 * pi
+      real (KIND=8), parameter ::      pi = 3.14159265359 !   pi
+      real (KIND=8), parameter ::      deg2rad = PI / 180.0  !  convert degrees to radians
+      real (KIND=8), parameter ::      pid2 = PI / 2.0 !   pi/2
+      real (KIND=8), parameter ::      rad2deg = 180.0 / PI !  convert radians to degrees
+      real (KIND=8), parameter ::      twopi = 2 * pi
 
 !  
 !***********************************************************************
